requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
--trusted-host artifactory.ing.net
--trusted-host files.pythonhosted.org
--trusted-host pypi.org

aiohttp==3.8.1
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
alembic==1.7.6
    # via apache-airflow
anyio==3.5.0
    # via httpcore
apache-airflow==2.2.3
    # via
    #   -r requirements.in
    #   apache-airflow-providers-postgres
apache-airflow-providers-ftp==2.0.1
    # via apache-airflow
apache-airflow-providers-http==2.0.3
    # via apache-airflow
apache-airflow-providers-imap==2.2.0
    # via apache-airflow
apache-airflow-providers-postgres==3.0.0
    # via -r requirements.in
apache-airflow-providers-sqlite==2.1.0
    # via apache-airflow
apispec[yaml]==3.3.2
    # via flask-appbuilder
argcomplete==1.12.3
    # via apache-airflow
async-timeout==4.0.2
    # via aiohttp
asyncio==3.4.3
    # via -r requirements.in
attrs==20.3.0
    # via
    #   aiohttp
    #   apache-airflow
    #   cattrs
    #   jsonschema
babel==2.9.1
    # via flask-babel
blinker==1.4
    # via apache-airflow
cattrs==1.5.0
    # via apache-airflow
certifi==2021.10.8
    # via
    #   httpx
    #   kaggle
    #   requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   httpx
    #   requests
click==7.1.2
    # via
    #   clickclick
    #   flask
    #   flask-appbuilder
clickclick==20.10.2
    # via apache-airflow
colorama==0.4.4
    # via
    #   flask-appbuilder
    #   rich
colorlog==5.0.1
    # via apache-airflow
commonmark==0.9.1
    # via rich
croniter==1.0.15
    # via apache-airflow
cryptography==36.0.1
    # via apache-airflow
defusedxml==0.7.1
    # via python3-openid
dill==0.3.4
    # via apache-airflow
dnspython==2.2.0
    # via email-validator
docutils==0.16
    # via
    #   apache-airflow
    #   python-daemon
email-validator==1.1.3
    # via flask-appbuilder
flask==1.1.4
    # via
    #   apache-airflow
    #   flask-appbuilder
    #   flask-babel
    #   flask-caching
    #   flask-jwt-extended
    #   flask-login
    #   flask-openid
    #   flask-sqlalchemy
    #   flask-wtf
flask-appbuilder==3.4.4
    # via apache-airflow
flask-babel==2.0.0
    # via flask-appbuilder
flask-caching==1.10.1
    # via apache-airflow
flask-jwt-extended==3.25.1
    # via flask-appbuilder
flask-login==0.4.1
    # via
    #   apache-airflow
    #   flask-appbuilder
flask-openid==1.3.0
    # via flask-appbuilder
flask-sqlalchemy==2.5.1
    # via flask-appbuilder
flask-wtf==0.14.3
    # via
    #   apache-airflow
    #   flask-appbuilder
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
graphviz==0.19.1
    # via apache-airflow
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via apache-airflow
h11==0.12.0
    # via httpcore
httpcore==0.13.7
    # via httpx
httpx==0.19.0
    # via apache-airflow
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
    #   rfc3986
    #   yarl
importlib-metadata==4.11.1
    # via markdown
inflection==0.5.1
    # via apache-airflow
iso8601==1.0.2
    # via apache-airflow
itsdangerous==1.1.0
    # via
    #   apache-airflow
    #   flask
    #   flask-wtf
jinja2==2.11.3
    # via
    #   apache-airflow
    #   flask
    #   flask-babel
    #   python-nvd3
    #   swagger-ui-bundle
jsonschema==3.2.0
    # via
    #   apache-airflow
    #   flask-appbuilder
    #   openapi-schema-validator
    #   openapi-spec-validator
kaggle==1.5.12
    # via -r requirements.in
lazy-object-proxy==1.7.1
    # via apache-airflow
lockfile==0.12.2
    # via
    #   apache-airflow
    #   python-daemon
mako==1.1.6
    # via alembic
markdown==3.3.6
    # via apache-airflow
markupsafe==2.1.0
    # via
    #   apache-airflow
    #   jinja2
    #   mako
    #   wtforms
marshmallow==3.14.1
    # via
    #   flask-appbuilder
    #   marshmallow-enum
    #   marshmallow-oneofschema
    #   marshmallow-sqlalchemy
marshmallow-enum==1.5.1
    # via flask-appbuilder
marshmallow-oneofschema==3.0.1
    # via apache-airflow
marshmallow-sqlalchemy==0.26.1
    # via flask-appbuilder
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.22.2
    # via pandas
openapi-schema-validator==0.2.3
    # via openapi-spec-validator
openapi-spec-validator==0.4.0
    # via apache-airflow
packaging==21.3
    # via apache-airflow
pandas==1.4.1
    # via -r requirements.in
pendulum==2.1.2
    # via apache-airflow
prison==0.2.1
    # via flask-appbuilder
psutil==5.9.0
    # via apache-airflow
psycopg==3.0.9
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via apache-airflow-providers-postgres
pycparser==2.21
    # via cffi
pygments==2.11.2
    # via
    #   apache-airflow
    #   rich
pyjwt==1.7.1
    # via
    #   apache-airflow
    #   flask-appbuilder
    #   flask-jwt-extended
pyparsing==3.0.7
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-daemon==2.3.0
    # via apache-airflow
python-dateutil==2.8.2
    # via
    #   apache-airflow
    #   croniter
    #   flask-appbuilder
    #   kaggle
    #   pandas
    #   pendulum
python-dotenv==0.19.2
    # via -r requirements.in
python-nvd3==0.15.0
    # via apache-airflow
python-slugify==4.0.1
    # via
    #   apache-airflow
    #   kaggle
    #   python-nvd3
python3-openid==3.2.0
    # via
    #   apache-airflow
    #   flask-openid
pytz==2021.3
    # via
    #   babel
    #   flask-babel
    #   pandas
pytzdata==2020.1
    # via pendulum
pyyaml==6.0
    # via
    #   apache-airflow
    #   apispec
    #   clickclick
    #   openapi-spec-validator
requests==2.27.1
    # via
    #   -r requirements.in
    #   apache-airflow-providers-http
    #   kaggle
rfc3986[idna2008]==1.5.0
    # via httpx
rich==11.2.0
    # via apache-airflow
setproctitle==1.2.2
    # via apache-airflow
six==1.16.0
    # via
    #   flask-jwt-extended
    #   jsonschema
    #   kaggle
    #   prison
    #   python-dateutil
    #   sqlalchemy-utils
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.31
    # via
    #   alembic
    #   apache-airflow
    #   flask-appbuilder
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
    #   sqlalchemy-jsonfield
    #   sqlalchemy-utils
sqlalchemy-jsonfield==1.0.0
    # via apache-airflow
sqlalchemy-utils==0.38.2
    # via flask-appbuilder
swagger-ui-bundle==0.0.9
    # via apache-airflow
tabulate==0.8.9
    # via apache-airflow
tenacity==8.0.1
    # via apache-airflow
termcolor==1.1.0
    # via apache-airflow
text-unidecode==1.3
    # via python-slugify
tqdm==4.62.3
    # via kaggle
unicodecsv==0.14.1
    # via apache-airflow
urllib3==1.26.8
    # via
    #   kaggle
    #   requests
werkzeug==1.0.1
    # via
    #   apache-airflow
    #   flask
    #   flask-jwt-extended
wtforms==2.3.3
    # via
    #   apache-airflow
    #   flask-appbuilder
    #   flask-wtf
yarl==1.7.2
    # via aiohttp
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
